Add files via upload
The Excel file is a cleaned and organised version of raw data that has been analyzed and transformed into interactive dashboards using Power BI and pivot tables. The primary focus of the analysis was to identify areas for optimising profit based on the region. The file includes key performance metrics such as sales, profit, and quantity, and can provide valuable insights for decision-making and strategic planning.